html,js added and fix some bugs

+++ fastapi_proj/app/schemas.py
@@ -10,4 +10,4 @@ class ServiceCalculator(BaseModel):
 
 class ServiceDate(BaseModel):
     date : datetime.datetime
-    days : int
+    days : float
